+++ lib/core/error/exceptions.dart
@@ -3,18 +3,48 @@ const String clientExceptionMessage = 'Client Exception';
 const String cacheExceptionMessage = 'Cache Exception';
 // const String formatExceptionMessage = 'Format Exception';
 
+Never throwException({
+  required String message,
+  required int code,
+  String? url,
+}) {
+  if (code >= 500) {
+    throw ServerException(
+      code: code,
+      message: message,
+      uri: url != null ? Uri.parse(url) : null,
+    );
+  } else if (code >= 400) {
+    throw ClientException(
+      code: code,
+      message: message,
+      uri: url != null ? Uri.parse(url) : null,
+    );
+  } else {
+    throw Exception(message);
+  }
+}
+
 class ServerException implements Exception {
   ServerException({
     this.code = 500,
     this.message = serverExceptionMessage,
+    this.uri,
   });
 
   final int code;
   final String message;
 
+  /// The URL of the HTTP request or response that failed.
+  final Uri? uri;
+
   @override
   String toString() {
-    return 'ServerException: $message';
+    if (uri != null) {
+      return 'ServerException: $message, uri=$uri';
+    } else {
+      return 'ServerException: $message';
+    }
   }
 }