+++ lib/features/auth/data/repositories/auth_repository_impl.dart
@@ -1,36 +1,38 @@
-import 'package:connectivity_plus/connectivity_plus.dart';
+import 'dart:io';
+
 import 'package:dartz/dartz.dart';
 import 'package:flutter_vtv/core/constants/typedef.dart';
 import 'package:flutter_vtv/core/helpers/secure_storage_helper.dart';
 import 'package:flutter_vtv/features/auth/domain/entities/auth_entity.dart';
+import 'package:jwt_decoder/jwt_decoder.dart';
 
+import '../../../../core/constants/constants.dart';
 import '../../../../core/error/exceptions.dart';
 import '../../../../core/error/failures.dart';
+import '../../domain/dto/register_params.dart';
 import '../../domain/repositories/auth_repository.dart';
 import '../data_sources/auth_data_source.dart';
 import '../models/auth_model.dart';
 
 class AuthRepositoryImpl implements AuthRepository {
-  AuthRepositoryImpl(this._authDataSource, this._connectivity, this._secureStorageHelper);
+  AuthRepositoryImpl(this._authDataSource, this._secureStorageHelper);
 
   final AuthDataSource _authDataSource;
-  final Connectivity _connectivity;
   final SecureStorageHelper _secureStorageHelper;
 
   @override
   FResult<AuthEntity> loginWithUsernameAndPassword(String username, String password) async {
     try {
-      final connectivityResult = await (_connectivity.checkConnectivity());
-      if (connectivityResult == ConnectivityResult.none) {
-        return const Left(ConnectionFailure(message: 'Không có kết nối mạng. Vui lòng kiểm tra lại.'));
-      }
-
       final model = await _authDataSource.loginWithUsernameAndPassword(username, password);
       return Right(model.toEntity());
+    } on SocketException {
+      return const Left(ServerFailure(message: kMsgNetworkError));
     } on ClientException catch (e) {
       return Left(ClientFailure(code: e.code, message: e.message));
     } on ServerException catch (e) {
       return Left(ServerFailure(code: e.code, message: e.message));
+    } catch (e) {
+      return Left(UnexpectedFailure(message: e.toString()));
     }
   }
 
@@ -43,7 +45,7 @@ class AuthRepositoryImpl implements AuthRepository {
     } on CacheException {
       return const Left(UnexpectedFailure(message: 'Lỗi lưu thông tin người dùng!'));
     } catch (e) {
-      return const Left(UnexpectedFailure(message: 'Lỗi không xác định! [cacheAuth]'));
+      return Left(UnexpectedFailure(message: e.toString()));
     }
   }
 
@@ -55,7 +57,7 @@ class AuthRepositoryImpl implements AuthRepository {
     } on CacheException catch (e) {
       return Left(CacheFailure(message: e.message));
     } catch (e) {
-      return const Left(UnexpectedFailure(message: 'Lỗi không xác định! [retrieveAuth]'));
+      return Left(UnexpectedFailure(message: e.toString()));
     }
   }
 
@@ -63,10 +65,12 @@ class AuthRepositoryImpl implements AuthRepository {
   FResultVoid logout(String refreshToken) async {
     try {
       return Right(await _authDataSource.disableRefreshToken(refreshToken));
+    } on SocketException {
+      return const Left(ServerFailure(message: kMsgNetworkError));
     } on ClientException catch (e) {
       return Left(ClientFailure(code: e.code, message: e.message));
     } catch (e) {
-      return const Left(UnexpectedFailure(message: 'Lỗi không xác định! [logout]'));
+      return Left(UnexpectedFailure(message: e.toString()));
     }
   }
 
@@ -77,7 +81,50 @@ class AuthRepositoryImpl implements AuthRepository {
     } on CacheException {
       return const Left(CacheFailure(message: 'Lỗi xóa thông tin người dùng!'));
     } catch (e) {
-      return const Left(UnexpectedFailure(message: 'Lỗi không xác định! [deleteAuth]'));
+      return Left(UnexpectedFailure(message: e.toString()));
+    }
+  }
+
+  @override
+  FResult<String> getAccessToken(String refreshToken) async {
+    try {
+      final newAccessToken = await _authDataSource.getAccessToken(refreshToken);
+      return Right(newAccessToken);
+    } on SocketException {
+      return const Left(ServerFailure(message: kMsgNetworkError));
+    } on ServerException catch (e) {
+      return Left(ServerFailure(code: e.code, message: e.message));
+    } catch (e) {
+      return Left(UnexpectedFailure(message: e.toString()));
+    }
+  }
+
+  /// Returns true if the token is valid, false if it is expired.
+  /// When some error occurs, it returns a [Failure].
+  @override
+  FResult<bool> isValidToken(String accessToken) async {
+    try {
+      return Right(JwtDecoder.isExpired(accessToken));
+    } on FormatException catch (e) {
+      return Left(UnexpectedFailure(message: e.message));
+    } catch (e) {
+      return Left(UnexpectedFailure(message: e.toString()));
+    }
+  }
+
+  @override
+  FResult<void> register(RegisterParams registerParams) async {
+    try {
+      await _authDataSource.register(registerParams);
+      return const Right(null);
+    } on SocketException {
+      return const Left(ServerFailure(message: kMsgNetworkError));
+    } on ClientException catch (e) {
+      return Left(ClientFailure(code: e.code, message: e.message));
+    } on ServerException catch (e) {
+      return Left(ServerFailure(code: e.code, message: e.message));
+    } catch (e) {
+      return Left(UnexpectedFailure(message: e.toString()));
     }
   }
 }