+++ lib/features/auth/domain/repositories/auth_repository.dart
@@ -1,10 +1,23 @@
 import 'package:flutter_vtv/core/constants/typedef.dart';
+import 'package:flutter_vtv/features/auth/domain/dto/register_params.dart';
 import 'package:flutter_vtv/features/auth/domain/entities/auth_entity.dart';
 
 abstract class AuthRepository {
+  // start app
   FResult<AuthEntity> retrieveAuth();
+
+  // login
   FResult<AuthEntity> loginWithUsernameAndPassword(String username, String password);
   FResultVoid cacheAuth(AuthEntity authEntity);
-  FResultVoid deleteAuth();
+
+  // logout
   FResultVoid logout(String refreshToken);
+  FResultVoid deleteAuth();
+
+  // register
+  FResultVoid register(RegisterParams registerParams);
+
+  // token expired
+  FResult<bool> isValidToken(String accessToken); // check token expired
+  FResult<String> getAccessToken(String refreshToken);
 }